+++ changelog.md
@@ -1,3 +1,9 @@
+### Version 0.2.0 (Unreleased) ###
+
+- Allow passing a source map as a string and anything with a `.toJSON()` method
+  (such as a `SourceMapGenerator`) as well as an object.
+
+
 ### Version 0.1.0 (2014-03-22) ###
 
 - Initial release.